Updates
* moved channel api to channel_endpoints.go
* struct updates
