+++ api/permissions.go
@@ -123,7 +123,7 @@ func computeBasePermissions(guild *Guild, member *GuildMember) Permission {
 	return permissions
 }
 
-func getOverwrite(channel *Channel, id *Snowflake, pType PermissionType) *Overwrite {
+func getOverwrite(channel *Channel, id *Snowflake, pType OverwriteType) *Overwrite {
 	for _, overwrite := range channel.PermissionOverwrites {
 		if overwrite.Type == pType && overwrite.ID == *id {
 			return overwrite
